test: combine invariant tests into main test file
